zfcpdump: Add prefix registers to dump header
Patchwork-id: 29916
O-Subject: [RHEL6.1 PATCH 09/14] [s390x] zfcpdump: Add prefix registers to dump
	header
Bugzilla: 633454

Description
-----------
With this patch the prefix registers of all online CPUs are
stored in the the zcore dump header. This allows dump analysis
tools to access the register information that is stored in the
prefix pages without using the System.map. This feature is
also needed as preparation for makedumpfile support on s390x.

Bugzilla
--------
BZ 633454
https://bugzilla.redhat.com/show_bug.cgi?id=633454

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.34
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=0cbde8ee60c883b246fb50dfed8bcd968239702b

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>