dasd: Add tunable default grace period for missing interrupts
Patchwork-id: 29918
O-Subject: [RHEL6.1 PATCH 12/14] [s390x] dasd: Add tunable default grace period
	for missing interrupts
Bugzilla: 633386

Description
-----------
This feature provides a user interface to specify the timeout for
missing interrupts for standard I/O operations.

Every I/O operation has a specific timeout value within the storage
server has to return a response to the operating system.  If the device
fails to respond to a submitted request within this timespan the request
will be canceled and retried.

The default value for this timeout was 300 seconds for standard ECKD and
FBA I/O operations and 50 seconds for DIAG I/O operations. This feature
provides a sysfs interface to allow users to change the timeout value.
Furthermore, for ECKD devices the timeout value provided from the
storage server is used as default instead of the generic 300 seconds.
FBA and DIAG devices will continue to use the default value of 300s and
50s respectively if the user has not specified a different timeout.

The timeout can be set individually for each device. The value specifies
the timeout in seconds and must be an integer in the range 1 to
40.000.000. The timeout value can be read and set through a new DASD
device attribute, 'expires', in sysfs.

Bugzilla
--------
BZ 633386
https://bugzilla.redhat.com/show_bug.cgi?id=633386

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.36
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=7c8faa86290c1a2607d6b768a0b874ec392a5c2a

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>