enable state transistions from OFFLINE to RUNNING
Patchwork-id: 29929
O-Subject: [PATCH RHEL 6.1]: scsi: enable state transistions from OFFLINE to
	RUNNING
Bugzilla: 643237
RH-Acked-by: Rob Evers <revers@redhat.com>
RH-Acked-by: David Milburn <dmilburn@redhat.com>

From: Mike Christie <mchristi@redhat.com>

This is for BZ: 643237.

In this commit:
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=5c10e63c943b4c67561ddc6bf61e01d4141f881f
we stopped allowing devices to from offline to running.

The problem is that with qla4xxx and all FC drivers a device might be
offlined due to a transport problem, then later the problem is resolved
automatically but the scsi devices are not usable by the kernel until a user
manually corrects the state.

Upstream I am working on fixing this in a way that allows the
transport classes to handle the problem if they know it is a
transport problem. My original attmept is here:
http://marc.info/?l=linux-scsi&m=128521870108931&w=2
http://marc.info/?l=linux-scsi&m=128521870208934&w=2
http://marc.info/?l=linux-scsi&m=128521870808945&w=2
http://marc.info/?l=linux-scsi&m=128521869608923&w=2
http://marc.info/?l=linux-scsi&m=128521871008948&w=2
http://marc.info/?l=linux-scsi&m=128521870408938&w=2
http://marc.info/?l=linux-scsi&m=128521870608941&w=2
I self nackd those patches upstream, because they were buggy
but am working on fixes that should work for all drivers.

In the end it seems we will need something like my patches
above, but they are invassive and need a good deal of testing.
Right now, I think we need a quick fix for this problem because
it adds a regression from RHEL5's behavior and is preventing
vendors from certifying on RHEL 6. So this patch just reverts
the patch that added the problem.

Note:
The patch I am reverting does fix a problem where IO will take a long
time (have to go through 5 retries instead of just 1 run) to fail if the
device bounces between offline and running, but I would consider that more
of a optimization and with the patch multipath and other apps are broken.

I have tested the patch and so has Netapp.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>