KVM: abstract kvm x86 mmu->n_free_mmu_pages
Patchwork-id: 29933
O-Subject: [RHEL6.1 1/4] KVM: abstract kvm x86 mmu->n_free_mmu_pages
Bugzilla: 632772
RH-Acked-by: Avi Kivity <avi@redhat.com>
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Gleb Natapov <gleb@redhat.com>

From: Dave Hansen <dave@linux.vnet.ibm.com>

"free" is a poor name for this value.  In this context, it means,
"the number of mmu pages which this kvm instance should be able to
allocate."  But "free" implies much more that the objects are there
and ready for use.  "available" is a much better description, especially
when you see how it is calculated.

In this patch, we abstract its use into a function.  We'll soon
replace the function's contents by calculating the value in a
different way.

All of the reads of n_free_mmu_pages are taken care of in this
patch.  The modification sites will be handled in a patch
later in the series.

Signed-off-by: Dave Hansen <dave@linux.vnet.ibm.com>
Signed-off-by: Tim Pepper <lnxninja@linux.vnet.ibm.com>
Signed-off-by: Avi Kivity <avi@redhat.com>
upstream commit e0df7b9f6cee43c01d6f4a8491bccfd410cb86e1
BZ: 632772

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>