fsck.gfs2 reported statfs error after gfs2_grow
Patchwork-id: 29965
O-Subject: [RHEL6.1 Patch] fsck.gfs2 reported statfs error after gfs2_grow -
	bz #661048
Bugzilla: 661048
RH-Acked-by: Steven Whitehouse <swhiteho@redhat.com>
RH-Acked-by: Abhijith Das <adas@redhat.com>

Hi,

This patch is for RHEL6.1 bug #661048, which is a crosswrite
from RHEL5 bug #660661.

When you do gfs2_grow it failed to take the very last
rgrp into account when adding up the new free space due
to an off-by-one error.  It was not reading the last
rgrp from the rindex because of a check for ">=" that
should have been ">".  Therefore, fsck.gfs2 was finding
(and fixing) the error with the system statfs file.

With the proper ">" check, the GFS2 kernel properly counts
all the new resource groups and fixes the statfs file
correctly, so fsck.gfs2 doesn't need to clean it up.

This patch has been submitted upstream and accepted into
the -nmw (development) git tree.  It was tested on RHEL5
on system roth-01, where I could easily recreate the failure,
and it was found to fix the problem.

Regards,

Bob Peterson
Red Hat GFS

Signed-off-by: Bob Peterson <rpeterso@redhat.com>
--
commit 4219faf5c4f36c6ff4bc7ab900b6837c86fceb24
Author: Bob Peterson <rpeterso@redhat.com>
Date:   Tue Dec 7 14:52:13 2010 -0600

    fsck.gfs2 reported statfs error after gfs2_grow

    When you do gfs2_grow it failed to take the very last
    rgrp into account when adding up the new free space due
    to an off-by-one error.  It was not reading the last
    rgrp from the rindex because of a check for ">=" that
    should have been ">".  Therefore, fsck.gfs2 was finding
    (and fixing) an error with the system statfs file.

    rhbz#661048

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>