dasd: Improve handling of stolen DASD reservation
Patchwork-id: 29972
O-Subject: [RHEL6.1 PATCH 3/8] [s390x] dasd: Improve handling of stolen DASD
	reservation
Bugzilla: 644942

Description
-----------
If a DASD device has been reserved by a Linux system, and later
this reservation is "stolen" by a second system by means of an
unconditional reserve, then the first system receives a
notification about this fact. With this patch such an event can
be either ignored, as before, or it can be used to let the device
fail all I/O request, so that the device will not block anymore.

Bugzilla
--------
BZ 644942
https://bugzilla.redhat.com/show_bug.cgi?id=644942

Upstream status of the patch
----------------------------
The patch is queued for kernel version 2.6.38.
http://148.100.100.18/cgi-bin/gitweb.cgi?p=linux-2.6.git;a=shortlog;h=refs/heads/features

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>