zcrypt: Introduce check for 4096 bit support
Patchwork-id: 29970
O-Subject: [RHEL6.1 PATCH 4/8] [s390x] zcrypt: Introduce check for 4096 bit
	support
Bugzilla: 633458

Description
-----------
Implemented an asm in the ap bus and made it accessible for the card
specific parts of the zcrypt driver. Thus when a cex3a is recognized
a check can be performed to dermine whether the card supports 4096 bit
RSA keys.

Bugzilla
--------
BZ 633458
https://bugzilla.redhat.com/show_bug.cgi?id=633458

Upstream status of the patch
----------------------------
The patch is queued for kernel version 2.6.38.
http://148.100.100.18/cgi-bin/gitweb.cgi?p=linux-2.6.git;a=shortlog;h=refs/heads/features

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>