zcrypt: cope with cca restriction of cex3
Patchwork-id: 29973
O-Subject: [RHEL6.1 PATCH 7/8] [s390x] zcrypt: cope with cca restriction of cex3
Bugzilla: 633458

Description
-----------
The cca on the crypto adapter has a restriction in the size of the
exponent if a key with a modulus bigger than 2048 bit is used. Thus
in that case we have to avoid that the crypto device driver thinks
the adapter is defect and sets it offline. Therfore a new member for
the zcrypt_device struct called max_exp_bit_length is introduced. This
will be set the first time the cca returns the error code function
not implemented. If this is done with an adapter twice it will return
-EINVAL.

Bugzilla
--------
BZ 633458
https://bugzilla.redhat.com/show_bug.cgi?id=633458

Upstream status of the patch
----------------------------
The patch is queued for kernel version 2.6.38.
http://148.100.100.18/cgi-bin/gitweb.cgi?p=linux-2.6.git;a=shortlog;h=refs/heads/features

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>