cifs: keep dentry reference in cifsFileInfo instead of inode reference
Patchwork-id: 30050
O-Subject: [RHEL6.1 PATCH 04/23] BZ#656461: cifs: keep dentry reference in
	cifsFileInfo instead of inode reference
Bugzilla: 656461
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

cifsFileInfo is a bit problematic. It contains a reference back to the
struct file itself. This makes it difficult for a cifsFileInfo to exist
without a corresponding struct file.

It would be better instead of the cifsFileInfo just held info pertaining
to the open file on the server instead without any back refrences to the
struct file. This would allow it to exist after the filp to which it was
originally attached was closed.

Much of the use of the file pointer in this struct is to get at the
dentry.  Begin divorcing the cifsFileInfo from the struct file by
keeping a reference to the dentry. Since the dentry will have a
reference to the inode, we can eliminate the "pInode" field too and
convert the igrab/iput to dget/dput.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Reviewed-by: Suresh Jayaraman <sjayaraman@suse.de>
Acked-by: Dave Kleikamp <shaggy@linux.vnet.ibm.com>
Signed-off-by: Steve French <sfrench@us.ibm.com>
Upstream commit: a5e18bc36e9e05ce0338d370a2ce4290910e43ea

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>