cifs: convert GlobalSMBSeslock from a rwlock to regular spinlock
Patchwork-id: 30041
O-Subject: [RHEL6.1 PATCH 15/23] BZ#656461: cifs: convert GlobalSMBSeslock from
	a rwlock to regular spinlock
Bugzilla: 656461
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

Convert this lock to a regular spinlock

A rwlock_t offers little value here. It's more expensive than a regular
spinlock unless you have a fairly large section of code that runs under
the read lock and can benefit from the concurrency.

Additionally, we need to ensure that the refcounting for files isn't
racy and to do that we need to lock areas that can increment it for
write. That means that the areas that can actually use a read_lock are
very few and relatively infrequently used.

While we're at it, change the name to something easier to type, and fix
a bug in find_writable_file. cifsFileInfo_put can sleep and shouldn't be
called while holding the lock.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Reviewed-by: Suresh Jayaraman <sjayaraman@suse.de>
Signed-off-by: Steve French <sfrench@us.ibm.com>
Upstream commit: 4477288a103631980750c86547d1fd54bfd2ba7d

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>