cifs: move close processing from cifs_close to cifsFileInfo_put
Patchwork-id: 30042
O-Subject: [RHEL6.1 PATCH 17/23] BZ#656461: [CIFS] move close processing from
	cifs_close to cifsFileInfo_put
Bugzilla: 656461
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>

From: Steve French <sfrench@us.ibm.com>

Now that it's feasible for a cifsFileInfo to outlive the filp under
which it was created, move the close processing into cifsFileInfo_put.

This means that the last user of the filehandle always does the actual
on the wire close call. This also allows us to get rid of the closePend
flag from cifsFileInfo. If we have an active reference to the file
then it's never going to have a close pending.

cifs_close is converted to simply put the filehandle.

Signed-off-by: Jeff Layton <jlayton@redhat.com>
Reviewed-by: Suresh Jayaraman <sjayaraman@suse.de>
Signed-off-by: Steve French <sfrench@us.ibm.com>
Upstream commit: cdff08e76612e53580139653403aedea979aa639

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>