kernel: virtualization aware cpu measurement
Patchwork-id: 30071
O-Subject: [RHEL6.1 PATCH 2/2] [s390x] kernel: virtualization aware cpu
	measurement
Bugzilla: 631483

Description
-----------
Use the SPP instruction to set a tag on entry to / exit of the
virtual machine context. This allows the cpu measurement facility
to distinguish the samples from the host and the different guests.

Bugzilla
--------
BZ 631483
https://bugzilla.redhat.com/show_bug.cgi?id=631483

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 2.6.35
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=cd3b70f5d4d82f85d1e1d6e822f38ae098cf7c72
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=215b3096371907e5d866bb219be7ef3d5ce6c083

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>