NFS: Ensure we return the dirent->d_type when it is known
Patchwork-id: 30171
O-Subject: [RHEL6.1 PATCH 15/27] NFS: Ensure we return the dirent->d_type when
	it is known
Bugzilla: 662782
RH-Acked-by: J. Bruce Fields <bfields@redhat.com>
RH-Acked-by: Jeff Layton <jlayton@redhat.com>

From: Trond Myklebust <Trond.Myklebust@netapp.com>

Store the dirent->d_type in the struct nfs_cache_array_entry so that we
can use it in getdents() calls.

This fixes a regression with the new readdir code.

Signed-off-by: Trond Myklebust <Trond.Myklebust@netapp.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>