allocate memory in khugepaged outside of mmap_sem write mode
Patchwork-id: 30379
O-Subject: Re: [RHEL6 PATCH] allocate memory in khugepaged outside of
	mmap_sem write mode
Bugzilla: 647849
RH-Acked-by: Rik van Riel <riel@redhat.com>
RH-Acked-by: Johannes Weiner <jweiner@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=647849

https://brewweb.devel.redhat.com/taskinfo?taskID=2979741

I found a silly local variable declaration hidden a missing
parameter (not sure why my build environment isn't spawning a warning
of uninitialized variable for this one, if it did it would have never
happened in the first place as I check for warnings in mm/*).

I already sent the update for Andrew too.

The incremental patch would be this one:

Here a replacement patch for RHEL6 that obsoletes the one in
Message-ID: <20101215193206.GA5638@random.random>:

============
Subject: allocate memory in khugepaged outside of mmap_sem write mode

From: Andrea Arcangeli <aarcange@redhat.com>

This tries to be more friendly to filesystem in userland, with userland
backends that allocate memory in the I/O paths and that could deadlock if
khugepaged holds the mmap_sem write mode of the userland backend while
allocating memory. Memory allocation may wait for writeback I/O completion from
the daemon that may be blocked in the mmap_sem read mode if a page fault
happens and the daemon wasn't using mlock for the memory required for the I/O
submission and completion.

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>