fix the return value of scsi_target_queue_read()
Patchwork-id: 30534
O-Subject: [RHEL 6.1 PATCH 1/2] SCSI: fix the return value of
	scsi_target_queue_read()
Bugzilla: 643236
RH-Acked-by: David Milburn <dmilburn@redhat.com>

From: Mike Christie <mchristi@redhat.com>

This is for bz: 643236.

>From the upstream mail:

It seems that zero should be returned if scsi_target_is_busy(starget) is
true, no matter if sdev is on the starved list.

[Info about issue it fixes in rhel from me]

This causes a problem where drivers will set the target->can_queue
and expect the scsi layer will not send more IO than that. Without
the patch the scsi layer will send more IO and will cause drivers
to use resources that they have in reserve for error handling
or internal IO on normal IO. If the error handler runs or
the driver needs to send a internal IO then we will not have
the resources to complete the operation and that can lead to
devices getting offlined or connections getting dropped.

This patch is not upstream yet. It looks like it got lost
because the original auther did not send the patch correctly
formatted. I resent the patch here:
http://marc.info/?l=linux-scsi&m=129253123525451&w=2

I have tested this with bnx2i and cxgb3i cards and confirmed
it fixes the overqueueing problem.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>