be2net: Avoid firmware update when interface is not open
Patchwork-id: 30549
O-Subject: [RHEL6.1 PATCH] be2net: Avoid firmware update when interface is not
	open.
Bugzilla: 651952
RH-Acked-by: David S. Miller <davem@redhat.com>
RH-Acked-by: Stefan Assmann <sassmann@redhat.com>
RH-Acked-by: John Linville <linville@redhat.com>

BZs:
#651952 - be2net: error when flashing firmware

Description:
Attempting a firmware update operation when interface is down could lead
to failure of operation. This fix fails the request if the interface is not
running.

Upstream commits:
d9efd2a be2net: Fix to avoid firmware update when interface is not open.

Signed-off-by: Ivan Vecera <ivecera@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>