oprofile: implement IBS cpuid feature detection
Patchwork-id: 30713
O-Subject: [RHEL 6.1 PATCH 02/55] oprofile/x86: implement IBS cpuid feature
	detection
Bugzilla: 647750
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Robert Richter <robert.richter@amd.com>

commit 64683da6643e8c6c93f1f99548399b08c029fd13 upstream.

This patch adds IBS feature detection using cpuid flags. An IBS
capability mask is introduced to test for certain IBS features. The
bit mask is the same as for IBS cpuid feature flags (Fn8000_001B_EAX),
but bit 0 is used to indicate the existence of IBS.

The patch also changes the handling of the IbsOpCntCtl bit (periodic
op counter count control). The oprofilefs file for this feature
(ibs_op/dispatched_ops) will be only exposed if the feature is
available, also the default for the bit is set to count clock cycles.

In general, the userland can detect the availability of a feature by
checking for the corresponding file in oprofilefs. If it exists, the
feature also exists. This may lead to a dynamic file layout depending
on the cpu type with that the userland has to deal with. Current
opcontrol is compatible.

Signed-off-by: Robert Richter <robert.richter@amd.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>