oprofile: convert oprofile from timer_hook to hrtimer
Patchwork-id: 30724
O-Subject: [RHEL 6.1 PATCH 07/55] oprofile: convert oprofile from timer_hook to
	hrtimer
Bugzilla: 647750
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>
RH-Acked-by: Don Zickus <dzickus@redhat.com>

From: Martin Schwidefsky <schwidefsky@de.ibm.com>

commit bc078e4eab65f11bbaeed380593ab8151b30d703 upstream.

Oprofile is currently broken on systems running with NOHZ enabled.
A maximum of 1 tick is accounted via the timer_hook if a cpu sleeps
for a longer period of time. This does bad things to the percentages
in the profiler output. To solve this problem convert oprofile to
use a restarting hrtimer instead of the timer_hook.

Signed-off-by: Martin Schwidefsky <schwidefsky@de.ibm.com>
Signed-off-by: Robert Richter <robert.richter@amd.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>