libfc: Return a valid return code in fc_fcp_pkt_abort()
Patchwork-id: 31101
O-Subject: [RHEL6.1 PATCH 53/55] libfc: Return a valid return code in
	fc_fcp_pkt_abort()
Bugzilla: 633915 663561
RH-Acked-by: Rob Evers <revers@redhat.com>
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>

From: Mike Christie <mchristi@redhat.com>

This is for bz 663561.

This is in the fcoe maintainer's tree. It is going to be
sent to the scsi maitainer for 2.6.38.

From: Hillf Danton <dhillf@gmail.com>

libfc: Return a valid return code in fc_fcp_pkt_abort()

    Here ticks_left is added to record the result of
    wait_for_completion_timeout().

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>