libfc: Cleanup return paths in fc_rport_error_retry
Patchwork-id: 31102
O-Subject: [RHEL6.1 PATCH 54/55] libfc: Cleanup return paths in
	fc_rport_error_retry
Bugzilla: 633915 663561
RH-Acked-by: Rob Evers <revers@redhat.com>
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>

From: Mike Christie <mchristi@redhat.com>

This is for bz 663561.

This is in the fcoe maintainer's tree. It is going to be
sent to the scsi maitainer for 2.6.38.

From: Hillf Danton <dhillf@gmail.com>

libfc: Cleanup return paths in fc_rport_error_retry

    This patch makes it so that we only have one call to
    fc_rport_error. This patch does not completely
    consolidate return statements, there is still one return
    used when not calling fc_rport_error, but alternative
    solutions made the code more confusing.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>