libfc: dereferencing ERR_PTR in fc_tm_done()
Patchwork-id: 31103
O-Subject: [RHEL6.1 PATCH 55/55] libfc: dereferencing ERR_PTR in fc_tm_done()
Bugzilla: 633915 663561
RH-Acked-by: Rob Evers <revers@redhat.com>
RH-Acked-by: Tomas Henzl <thenzl@redhat.com>

From: Mike Christie <mchristi@redhat.com>

This is for bz 663561.

This is in the fcoe maintainer's tree. It is going to be
sent to the scsi maitainer for 2.6.38.

From: Dan Carpenter <error27@gmail.com>

In open-fcoe.org's patchworks, this is good bug fix:

http://www.open-fcoe.org/patchwork/patch/1554/

If we goto out, then it tries to call kfree_skb() on an ERR_PTR which
will oops.  Just return directly.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>