fix WinXP BSOD when boot up with -cpu Penryn
Patchwork-id: 31176
O-Subject: [PATCH RHEL6.1] Bug 635539 - WinXP BSOD when boot up with -cpu Penryn
Bugzilla: 635539
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Prarit Bhargava <prarit@redhat.com>

https://bugzilla.redhat.com/show_bug.cgi?id=635539

A correction to Intel cpu model CPUID data caused winxp-64
to BSOD when booted with a Penryn model.  This was traced
to the CPUID "model" field correction from 6 -> 23 (as is
proper for this class of cpu).  Only in this case does the
problem surface.

The cause for this failure is winxp accessing the BBL_CR_CTL3
MSR which is unsupported by current kvm, appears to be a
legacy MSR undocumented but existing in current silicon, and is
apparently carried forward in MSR space to accommodate vintage
code as here.  It is not yet conclusive whether this MSR
implements any of its legacy functionality (unlikely) or is
just an ornamental dud for compatibility.  I found no
mention of this MSR in contemporary cpu documentation and
after considerable search have found only this partial intel
doc describing the encoding:

http://cleo.tlv.redhat.com/qumrawiki/KVM/CPUModels?action=AttachFile&do=get&target=old-MSRs-intel-part4.pdf

.. agreeing with the functional behavior of other bootloader/kernel
code I've examined which accesses this MSR.  Regrettably winxp-64
appears to be setting bit #19 called out as "reserved" in
the above document.  So to emulate this MSR, kvm msr get will
provide the equivalent mock data and kvm msr write will simply
toss the guest passed data without interpretation.

Brew Build:  https://brewweb.devel.redhat.com/taskinfo?taskID=3013865

Upstream Status:  Requested clarification from Intel.  It is
probably fine pushed as-is given the narrow window of this
problem but having concrete vendor input would offset the
otherwise lack of documentation and help justify the approach.

Signed-off-by: john cooper <john.cooper@redhat.com>