cxgbi: set ulpmode only if digest is on
Patchwork-id: 31213
O-Subject: [RHEL6.1 PATCH 10/10] cxgbi: set ulpmode only if digest is on
Bugzilla: 636291

From: Mike Christie <mchristi@redhat.com>

This is for bz 636291.

This is not upstream yet. It was sent here:
http://marc.info/?l=linux-scsi&m=129444035725243&w=2

From: Karen Xie <kxie@chelsio.com>

There is no need to set ulpmode on the tx skbs if no digest is enabled.

Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>