vt: fix issue when fbcon wants to takeover a second time
Patchwork-id: 31263
O-Subject: [PATCH 08/11] [RHEL6.1 PATCH] vt: fix issue when fbcon wants to
	takeover a second time.
Bugzilla: 667281
RH-Acked-by: Adam Jackson <ajax@redhat.com>

bz: 667281
upstream: accepted by maintainer for 2.6.38-rc2

With framebuffer handover and multiple GPUs, we get into a
position where the fbcon unbinds the vesafb framebuffer for GPU 1,
but we still have a radeon framebuffer bound from GPU 0, so
we don't unregister the console driver. Then when we tried to bind
the new radeon framebuffer for GPU1 we never get to the bind
call as we fail due to the console being registered already.

This changes the return value to -EBUSY when the driver is
already registered and continues to bind for -EBUSY.

Signed-off-by: Dave Airlie <airlied@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>