BNX2I: Fixed the TCP graceful termination initiation
Patchwork-id: 31435
O-Subject: [RHEL 6.1 PATCH 02/24] BNX2I: Fixed the TCP graceful termination
	initiation
Bugzilla: 635894
RH-Acked-by: John Linville <linville@redhat.com>

From: Eddie Wai <eddie.wai@broadcom.com>

>From upstream commit 2eefb20dbf3032da1ad111c1ce178f899bc4859a.

In compliance to RFC793, a TCP graceful termination will be used
instead of an abortive termination for the case where the remote
has initiated the close of the connection.
Additionally, a TCP abortive termination will be used to close the
connection when a logout response is not received in time after a
logout request has been initiated.

Signed-off-by: Eddie Wai <eddie.wai@broadcom.com>
Reviewed-by: Michael Chan <mchan@broadcom.com>
Reviewed-by: Benjamin Li <benli@broadcom.com>
Acked-by: Anil Veerabhadrappa <anilgv@broadcom.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>