BNX2I: Allow to abort the connection if connect request times out
Patchwork-id: 31440
O-Subject: [RHEL 6.1 PATCH 11/24] BNX2I: Allow to abort the connection if
	connect request times out
Bugzilla: 635894
RH-Acked-by: John Linville <linville@redhat.com>

From: Eddie Wai <eddie.wai@broadcom.com>

>From upstream commit 252e44805bcfff5de2e05dd126f69a18e0b760f0.

In the situation where the connect completion response arrives after
the connect request has already timed out, the connection was not being
aborted but only the resource was being freed.  This creates a problem
for 5771X (10g) as the chip flags this with an assertion.

This change will properly aborts the connection before freeing the
resource.

Signed-off-by: Eddie Wai <eddie.wai@broadcom.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>