BNX2I: Added return code check for chip kwqe submission request
Patchwork-id: 31444
O-Subject: [RHEL 6.1 PATCH 15/24] BNX2I: Added return code check for chip kwqe
	submission request
Bugzilla: 635894
RH-Acked-by: John Linville <linville@redhat.com>

From: Eddie Wai <eddie.wai@broadcom.com>

>From upstream commit bee348770dcaeafa57bbd32129ce0cb16a4342d6.

Added the handling for cases when a chip request is made to the
CNIC module but the hardware is not ready to accept.  This would
lead to many unnecessary wait timeouts.
This code adds check in the connect establishment and destruction
path.

Signed-off-by: Eddie Wai <eddie.wai@broadcom.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>