xen: don't bother to stop other cpus on shutdown/reboot
Patchwork-id: 31482
O-Subject: [RHEL6 PATCH 1/2] xen: don't bother to stop other cpus on
	shutdown/reboot
Bugzilla: 667340
RH-Acked-by: Don Dutile <ddutile@redhat.com>
RH-Acked-by: Andrew Jones <drjones@redhat.com>
RH-Acked-by: Laszlo Ersek <lersek@redhat.com>
RH-Acked-by: Vivek Goyal <vgoyal@redhat.com>

Xen will shoot all the VCPUs when we do a shutdown hypercall, so there's
no need to do it manually.

In any case it will fail because all the IPI irqs have been pulled
down by this point, so the cross-CPU calls will simply hang forever.

With the next change the function calls will be synchronously waited
for, so this would become apparent leading to a hang on shutdown on
multi-VCPU guests.

Signed-off-by: Jeremy Fitzhardinge <jeremy.fitzhardinge@citrix.com>
Signed-off-by: Paolo Bonzini <pbonzini@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>