Btrfs: stop trying to shrink delalloc if there are no inodes to reclaim
Patchwork-id: 31626
O-Subject: [RHEL6.1 PATCH 02/81] Btrfs: stop trying to shrink delalloc if there
	are no inodes to reclaim
Bugzilla: 663749
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

In very severe ENOSPC cases we can run out of inodes to do delalloc on, which
means we'll just keep looping trying to shrink delalloc.  Instead, if we fail to
shrink delalloc 3 times in a row break out since we're not likely to make any
progress.  Tested this with a 100mb fs an xfstests test 13.  Before the patch it
would hang the box, with the patch we get -ENOSPC like we should.  Thanks,

Signed-off-by: Josef Bacik <josef@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>