Btrfs: fix df regression
Patchwork-id: 31625
O-Subject: [RHEL6.1 PATCH 04/81] Btrfs: fix df regression
Bugzilla: 663749
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

The new ENOSPC stuff breaks out the raid types which breaks the way we were
reporting df to the system.  This fixes it back so that Available is the total
space available to data and used is the actual bytes used by the filesystem.
This means that Available is Total - data used - all of the metadata space.
Thanks,

Signed-off-by: Josef Bacik <josef@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>