Btrfs: don't allocate chunks as aggressively
Patchwork-id: 31629
O-Subject: [RHEL6.1 PATCH 07/81] Btrfs: don't allocate chunks as aggressively
Bugzilla: 663749
RH-Acked-by: Eric Sandeen <sandeen@redhat.com>

Because the ENOSPC code over reserves super aggressively we end up allocating
chunks way more often than we should.  For example with my fs_mark tests on a
2gb fs I can end up reserved 1gb just for metadata, when only 34mb of that is
being used.  So instead check to see if the amount of space actually used is
less than 30% of the total space, and if so don't allocate a chunk, but only if
we have at least 256mb of free space to make sure we don't put too much pressure
on free space.

Signed-off-by: Josef Bacik <josef@redhat.com>
Signed-off-by: Aristeu Rozanski <arozansk@redhat.com>